+++ packages/accounts-controller/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `clearState` method to reset controller state to defaults, exposed as `AccountsController:clearState`
+  - Corresponding action type `AccountsControllerClearStateAction` is available as well.
+
 ## [39.0.5]
 
 ### Changed